# openapi.yaml
openapi: 3.1.0
info:
  title: Colette API
  description: Public REST API for the Colette app. Supports OAuth 2.0 and API key authentication.
  license:
    name: MIT
  version: 0.1.0
servers:
- url: /api
paths:
...
  /bookmarks:
    get:
      tags:
      - Bookmarks
      description: List user bookmarks
      operationId: listBookmarks
      parameters:
      - name: collectionId
        in: query
        description: Filter by the ID of a collection whose filters may apply to the bookmark
        required: false
        schema:
          type: string
          format: uuid
      - name: filterByTags
        in: query
        description: Whether to filter by tags linked to the bookmark
        required: false
        schema:
          type: boolean
      - name: tag[]
        in: query
        description: Filter by the IDs of the tags linked to the bookmark
        required: false
        schema:
          type: array
          items:
            type: string
            format: uuid
      - name: cursor
        in: query
        description: Pagination cursor
        required: false
        schema:
          type: string
      - name: withTags
        in: query
        description: Whether to include the tags linked to the bookmark
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Paginated list of bookmarks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paginated_BookmarkDetails'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - Bookmarks
      description: Add a bookmark
      operationId: createBookmark
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkCreate'
        required: true
      responses:
        '201':
          description: Created bookmark
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bookmark'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Bookmark already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /bookmarks/{id}:
    get:
      tags:
      - Bookmarks
      description: Get a bookmark by ID
      operationId: getBookmark
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      - name: withTags
        in: query
        description: Whether to include the tags linked to the bookmark
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Bookmark by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkDetails'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Bookmark not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - Bookmarks
      description: Delete a bookmark by ID
      operationId: deleteBookmark
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Successfully deleted bookmark
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Bookmark not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - Bookmarks
      description: Update a bookmark by ID
      operationId: updateBookmark
      parameters:
      - name: id
        in: path
        description: Unique identifier of the resource
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkUpdate'
        required: true
      responses:
        '200':
          description: Updated bookmark
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bookmark'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: User not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Bookmark not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /bookmarks/scrape:
    post:
      tags:
      - Bookmarks
      description: Scrape bookmark from a webpage
      operationId: scrapeBookmark
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkScrape'
        required: true
      responses:
        '201':
          description: Scraped bookmark
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkScraped'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '502':
          description: Failed to fetch data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /feeds/detect:
    post:
      tags:
      - Feeds
      description: Detects web feeds on a page
      operationId: detectFeeds
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedDetect'
        required: true
      responses:
        '201':
          description: List of detected feeds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeedDetected'
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '502':
          description: Failed to fetch data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        default:
          description: Unknown error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        message:
          type: string
    ApiErrorCode:
      type: string
      enum:
      - NOT_AUTHENTICATED
      - NOT_AUTHORIZED
      - NOT_FOUND
      - ALREADY_EXISTS
      - VALIDATION
      - BAD_GATEWAY
      - UNKNOWN
...
    Bookmark:
      type: object
      description: Bookmark to a webpage
      required:
      - id
      - link
      - title
      - thumbnailUrl
      - publishedAt
      - author
      - archivedUrl
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the bookmark
        link:
          type: string
          format: uri
          description: URL of the webpage the bookmark links to
        title:
          type: string
          description: Title of the bookmark
        thumbnailUrl:
          type:
          - string
          - 'null'
          format: uri
          description: Thumbnail URL of the bookmark
        publishedAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Timestamp at which the bookmark was published
        author:
          type:
          - string
          - 'null'
          description: Author of the bookmark
        archivedUrl:
          type:
          - string
          - 'null'
          format: uri
          description: URL of the archived version of the bookmark's thumbnail
        createdAt:
          type: string
          format: date-time
          description: Timestamp at which the bookmark was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp at which the bookmark was modified
    BookmarkCreate:
      type: object
      description: Data to create a new bookmark
      required:
      - url
      - title
      properties:
        url:
          type: string
          format: uri
          description: URL of the webpage the bookmark links to
        title:
          type: string
          description: Human-readable name for the new bookmark, cannot be empty
          minLength: 1
        thumbnailUrl:
          type:
          - string
          - 'null'
          format: uri
          description: Thumbnail URL of the new bookmark, will be archived
        publishedAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Timestamp at which the bookmark was published
        author:
          type:
          - string
          - 'null'
          description: Author for the new bookmark, cannot be empty
          minLength: 1
...
    BookmarkDetails:
      type: object
      description: Extended details of a bookmark
      required:
      - bookmark
      properties:
        bookmark:
          $ref: '#/components/schemas/Bookmark'
          description: Bookmark itself, always present
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          description: Linked tags, present if requested
...
    BookmarkScrape:
      type: object
      description: Data to scrape a bookmark using
      required:
      - url
      properties:
        url:
          type: string
          format: uri
          description: URL of a webpage to scrape
    BookmarkScraped:
      type: object
      description: Scraped bookmark
      required:
      - link
      - title
      - thumbnailUrl
      - publishedAt
      - author
      properties:
        link:
          type: string
          format: uri
          description: URL of the webpage of the scraped bookmark
        title:
          type: string
          description: Title of the scraped bookmark
        thumbnailUrl:
          type:
          - string
          - 'null'
          format: uri
          description: Thumbnail URL of the scraped bookmark
        publishedAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Timestamp at which the scraped bookmark was published
        author:
          type:
          - string
          - 'null'
          description: Author of the scraped bookmark
...
    BookmarkUpdate:
      type: object
      description: Updates to make to an existing bookmark
      properties:
        title:
          type: string
          description: Human-readable name for the bookmark to update, cannot be empty
          minLength: 1
        thumbnailUrl:
          type:
          - string
          - 'null'
          format: uri
          description: Thumbnail URL of the bookmark to update, will be archived
        publishedAt:
          type:
          - string
          - 'null'
          format: date-time
          description: Timestamp at which the bookmark was published
        author:
          type:
          - string
          - 'null'
          description: Author of the bookmark to update, cannot be empty
          minLength: 1
...
    FeedDetect:
      type: object
      description: Data to detect RSS feeds using
      required:
      - url
      properties:
        url:
          type: string
          format: uri
          description: URL of a webpage to detect RSS feeds on
    FeedDetected:
      type: object
      description: Detected RSS feed
      required:
      - url
      - title
      properties:
        url:
          type: string
          format: uri
          description: URL of the detected RSS feed
        title:
          type: string
          description: Title of the detected RSS feed
...
    Paginated_BookmarkDetails:
      type: object
      description: Paginated list of results
      required:
      - data
      properties:
        data:
          type: array
          items:
            type: object
            description: Extended details of a bookmark
            required:
            - bookmark
            properties:
              bookmark:
                $ref: '#/components/schemas/Bookmark'
                description: Bookmark itself, always present
              tags:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
                description: Linked tags, present if requested
          description: Current set of results
        cursor:
          type: string
          description: Pagination cursor, only present if more results are available
...
    Tag:
      type: object
      description: Tag that can be attached to subscriptions and bookmarks
      required:
      - id
      - title
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the tag
        title:
          type: string
          description: Human-readable name of the tag, unique per user
        createdAt:
          type: string
          format: date-time
          description: Timestamp at which the tag was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp at which the tag was last modified
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []